# Improve performance when huge connections arrive within a short period