--- a/src/Servers/Kestrel/Transport.Sockets/src/SocketConnectionListener.cs
+++ b/src/Servers/Kestrel/Transport.Sockets/src/SocketConnectionListener.cs
@@ -86,7 +86,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Sockets
 
             EndPoint = listenSocket.LocalEndPoint;
 
-            listenSocket.Listen(512);
+            listenSocket.Listen(int.MaxValue);
 
             _listenSocket = listenSocket;
         }