@anurse 2019-11-05
Closing as this conflicts with the Backlog option we added. Also we certainly wouldn't just increase the backlog to `int.MaxValue`, it would need to be configurable (which is what we've done).