@halter73 2019-11-05
@303248153 This is now configurable via SocketTransportOptions and LibuvTransportOptions.

https://github.com/aspnet/AspNetCore/pull/15111